LayoutTests/imported/w3c/web-platform-tests/fetch/api/request/request-consume-empty.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Request consume empty bodies</title>
    <meta name="help" href="https://fetch.spec.whatwg.org/#request">
    <meta name="help" href="https://fetch.spec.whatwg.org/#body-mixin">
    <meta name="author" title="Canon Research France" href="https://www.crf.canon.fr">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
    function checkBodyText(request) {
      return request.text().then( function(bodyAsText) {
        assert_equals(bodyAsText, "", "Resolved value should be empty");
        assert_false(request.bodyUsed);
      });
    }

    function checkBodyBlob(request) {
      return request.blob().then(function(bodyAsBlob) {
        var promise = new Promise(function (resolve, reject) {
          var reader = new FileReader();
          reader.onload = function(evt) {
            resolve(reader.result)
          };
          reader.onerror = function() {
            reject("Blob's reader failed");
          };
          reader.readAsText(bodyAsBlob);
        });
        return promise.then(function(body) {
          assert_equals(body, "", "Resolved value should be empty");
          assert_false(request.bodyUsed);
        });
      });
    }

    function checkBodyArrayBuffer(request) {
      return request.arrayBuffer().then( function(bodyAsArrayBuffer) {
        assert_equals(bodyAsArrayBuffer.byteLength, 0, "Resolved value should be empty");
        assert_false(request.bodyUsed);
      });
    }

    function checkBodyJSON(request) {
      return request.json().then(
        (bodyAsJSON) => {
          assert_unreached("JSON parsing should fail");
        },
        () => {
          assert_false(request.bodyUsed);
        });
    }

    function checkBodyFormData(request) {
      return request.formData().then(function(bodyAsFormData) {
        assert_true(bodyAsFormData instanceof FormData, "Should receive a FormData");
        assert_false(request.bodyUsed);
     });
    }

    function checkRequestBody(bodyType, checkFunction) {
      promise_test(function(test) {
        var request = new Request("", {"method": "POST"});
        assert_false(request.bodyUsed);
        return checkFunction(request);
      }, "Consume request's body as " + bodyType);
    }

    var formData = new FormData();
    checkRequestBody("text", checkBodyText);
    checkRequestBody("blob", checkBodyBlob);
    checkRequestBody("arrayBuffer", checkBodyArrayBuffer);
    checkRequestBody("json", checkBodyJSON);
    checkRequestBody("formData", checkBodyFormData);

    </script>
  </body>
</html>
